# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Business Management Platform",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register/": {
      "post": {
        "tags": [
          "Аутентификация"
        ],
        "summary": "Регистрация пользователя",
        "description": "Регистрирует пользователя, назначая ему по умолчанию роль \"user\"\nПользователь существует и активен - 400 Bad Request\nРоль user не найдена - 500, ошибка на стороне сервера",
        "operationId": "register_user_api_v1_auth_register__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "400": {
            "description": "Пользователь уже существует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "Роль не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/restore/": {
      "patch": {
        "tags": [
          "Аутентификация"
        ],
        "summary": "Восстановление пользователя",
        "description": "Восстанавливает активность пользователя после 'мягкого' удаления\nПользователь уже активен - 409 Conflict\nНеправильные креды - 401 Unauthorized",
        "operationId": "restore_user_api_v1_auth_restore__patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserChangeStatus"
                }
              }
            }
          },
          "401": {
            "description": "Неверные данные учетной записи или пользователь не существует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "Пользователь уже активен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/": {
      "post": {
        "tags": [
          "Аутентификация"
        ],
        "summary": "Вход в систему",
        "description": "Аутентификация пользователя\nНеправильные креды - 401 Unauthorized\nПользователь есть, но неактивен - 403 Forbidden",
        "operationId": "login_api_v1_auth_login__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Неверные данные учетной записи или пользователь не существует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "Пользователь не активен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout/": {
      "post": {
        "tags": [
          "Аутентификация"
        ],
        "summary": "Выход из системы",
        "description": "Выход пользователя из системы\nПри логауте токен добавляется в блэклист",
        "operationId": "logout_api_v1_auth_logout__post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponseSchema": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponseSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "RoleRead",
        "description": "Схема для возврата модели роли клиенту"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token",
        "description": "Схема модели JWT-токена"
      },
      "UserChangeStatus": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "type": "object",
        "required": [
          "message",
          "user"
        ],
        "title": "UserChangeStatus",
        "description": "Схема для смены статуса пользователя (is_active)"
      },
      "UserLogin": {
        "properties": {
          "username": {
            "type": "string",
            "format": "email",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLogin",
        "description": "Схема для аутентификации пользователя"
      },
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "user@example.com"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Иван"
            ]
          },
          "surname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surname",
            "examples": [
              "Иванович"
            ]
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name",
            "examples": [
              "Иванов"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "role": {
            "$ref": "#/components/schemas/RoleRead"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "id",
          "is_active",
          "role"
        ],
        "title": "UserRead",
        "description": "Схема для возврата клиенту"
      },
      "UserRegister": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "user@example.com"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Иван"
            ]
          },
          "surname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surname",
            "examples": [
              "Иванович"
            ]
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name",
            "examples": [
              "Иванов"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 3,
            "title": "Password",
            "examples": [
              "secret_password"
            ]
          },
          "repeat_password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 3,
            "title": "Repeat Password",
            "examples": [
              "secret_password"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "name",
          "password",
          "repeat_password"
        ],
        "title": "UserRegister",
        "description": "Схема с полями для регистрации и валидацией пароля"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}